+++ plugins/flexdmd/actors/Frame.cpp
@@ -9,14 +9,14 @@ void Frame::Draw(Flex::SurfaceGraphics* pGraphics)
    if (GetVisible()) {
       if (m_fill) {
          pGraphics->SetColor(m_fillColor);
-         pGraphics->FillRectangle({ (int)(GetX() + m_thickness), (int)(GetY() + m_thickness), (int)(GetWidth() - 2 * m_thickness), (int)(GetHeight() - 2 * m_thickness) });
+         pGraphics->FillRectangle({ (int)(GetX() + (float)m_thickness), (int)(GetY() + (float)m_thickness), GetWidth() - 2 * m_thickness, GetHeight() - 2 * m_thickness });
       }
       if (m_thickness > 0) {
          pGraphics->SetColor(m_borderColor);
-         pGraphics->FillRectangle({ (int)GetX(), (int)GetY(), (int)GetWidth(), m_thickness });
-         pGraphics->FillRectangle({ (int)GetX(), (int)(GetY() + GetHeight() - m_thickness), (int)GetWidth(), m_thickness });
-         pGraphics->FillRectangle({ (int)GetX(), (int)(GetY() + m_thickness), m_thickness, (int)(GetHeight() - 2 * m_thickness) });
-         pGraphics->FillRectangle({ (int)(GetX() + GetWidth() - m_thickness), (int)(GetY() + m_thickness), m_thickness, (int)(GetHeight() - 2 * m_thickness) });
+         pGraphics->FillRectangle({ (int)GetX(), (int)GetY(), GetWidth(), m_thickness });
+         pGraphics->FillRectangle({ (int)GetX(), (int)(GetY() + (float)(GetHeight() - m_thickness)), GetWidth(), m_thickness });
+         pGraphics->FillRectangle({ (int)GetX(), (int)(GetY() + (float)m_thickness), m_thickness, GetHeight() - 2 * m_thickness });
+         pGraphics->FillRectangle({ (int)(GetX() + (float)(GetWidth() - m_thickness)), (int)(GetY() + (float)m_thickness), m_thickness, GetHeight() - 2 * m_thickness });
       }
    }
 }
