+++ plugins/flexdmd/actors/GIFImage.cpp
@@ -20,8 +20,8 @@ GIFImage* GIFImage::Create(FlexDMD* pFlexDMD, AssetManager* pAssetManager, const
    pImage->m_pSrc = pSrc;
    pImage->m_pAssetManager = pAssetManager;
    pImage->m_pBitmap = pBitmap;
-   pImage->SetPrefWidth(static_cast<float>(pBitmap->GetWidth()));
-   pImage->SetPrefHeight(static_cast<float>(pBitmap->GetHeight()));
+   pImage->SetPrefWidth(pBitmap->GetWidth());
+   pImage->SetPrefHeight(pBitmap->GetHeight());
    pImage->Rewind();
    pImage->Pack();
    pImage->m_pBitmap = nullptr;
