+++ plugins/flexdmd/actors/Image.cpp
@@ -24,8 +24,8 @@ Image* Image::Create(FlexDMD* pFlexDMD, AssetManager* pAssetManager, const strin
    pImage->m_pSrc = pSrc;
    pImage->m_pAssetManager = pAssetManager;
    pImage->m_pBitmap = pBitmap;
-   pImage->SetPrefWidth(static_cast<float>(pBitmap->GetWidth()));
-   pImage->SetPrefHeight(static_cast<float>(pBitmap->GetHeight()));
+   pImage->SetPrefWidth(pBitmap->GetWidth());
+   pImage->SetPrefHeight(pBitmap->GetHeight());
    pImage->Pack();
    pImage->m_pBitmap = nullptr;
