+++ plugins/flexdmd/resources/Font.h
@@ -16,9 +16,9 @@ public:
 
    PSC_IMPLEMENT_REFCOUNT()
 
-   void DrawCharacter(Flex::SurfaceGraphics* pGraphics, char character, char previousCharacter, float& x, float& y);
+   void DrawCharacter(Flex::SurfaceGraphics* pGraphics, char character, char previousCharacter, int& x, int& y);
    SDL_Rect MeasureFont(const string& text);
-   void DrawText_(Flex::SurfaceGraphics* pGraphics, float x, float y, const string& text);
+   void DrawText_(Flex::SurfaceGraphics* pGraphics, int x, int y, const string& text);
    BitmapFont* GetBitmapFont() const { return m_pBitmapFont; }
 
 private:
