+++ plugins/flexdmd/scenes/SingleLineScene.cpp
@@ -25,9 +25,9 @@ void SingleLineScene::Begin()
    BackgroundScene::Begin();
    if (m_scroll) {
       // UltraDMD moves text by 1.2 pixel per frame (no delta time) and seems to render based on a frame rate at 60FPS.
-      float offset = -(GetWidth() - m_pText->GetWidth()) / 2.0f;
-      m_scrollX = offset + GetWidth();
-      AddTween(m_scrollX, offset - m_pText->GetWidth(), (GetWidth() + m_pText->GetWidth()) / (60.0f * 1.2f), ([this](float newValue) {
+      float offset = (float)-(GetWidth() - m_pText->GetWidth()) / 2.0f;
+      m_scrollX = offset + (float)GetWidth();
+      AddTween(m_scrollX, offset - (float)m_pText->GetWidth(), (float)(GetWidth() + m_pText->GetWidth()) / (60.0f * 1.2f), ([this](float newValue) {
          this->m_scrollX = newValue;
          return false;
       }));
@@ -37,7 +37,7 @@ void SingleLineScene::Begin()
 void SingleLineScene::Update(float delta)
 {
    BackgroundScene::Update(delta);
-   m_pText->SetPosition((GetWidth() - m_pText->GetWidth()) / 2.0f + m_scrollX, (GetHeight() - m_pText->GetHeight()) / 2.0f);
+   m_pText->SetPosition((float)(GetWidth() - m_pText->GetWidth()) / 2.0f + m_scrollX, (float)(GetHeight() - m_pText->GetHeight()) / 2.0f);
 }
 
 }
